[feature]Created the parser package and implemented the basics of the parser
started chapter 2 and read until page 43. in this part we implemented the basics of a parser by parsing a variable definition statement known as Let statement in Monkey lang. we also added errors detection which checks for all the errors and then returns them all at once
TODO: One thing that can be improved is the error detection which can show the line number and the character